# Conflicts:
#	contracts/governance/GovernorAlpha.sol
#	contracts/governance/Staking/Checkpoints.sol
#	contracts/governance/Staking/Staking.sol

+++ contracts/governance/Vesting/IVesting.sol
@@ -1,8 +1,11 @@
 pragma solidity ^0.5.17;
 
 interface IVesting {
-    function duration() external returns (uint);
-    function endDate() external returns (uint);
-    function stakeTokens(uint amount) external;
-    function governanceWithdrawTokens(address receiver) external;
+	function duration() external returns (uint256);
+
+	function endDate() external returns (uint256);
+
+	function stakeTokens(uint256 amount) external;
+
+	function governanceWithdrawTokens(address receiver) external;
 }
